requirements.txt
pytest
requests
python-dotenv
beautifulsoup4
